Make operations easier to find
These aren’t really models, so I took my lead from the module they were
namespaced under and moved them to their own subdirectory.

TODO: rename them to prevent confusion with other classes in different
namespaces (GovPay API, for example).
